Final commit
The problem in the README.md file was the requests to an API that returned my status. It turns out that the request does not accept spaces between words, which was the value I was sending through the custom_title property.

After replacing the spaces with percent-encoding for special characters, the problem was solved.